Rearrange compile tests to share the same environment
